+++ pineappl_cli/src/sum.rs
@@ -2,6 +2,7 @@ use super::helpers::{self, Subcommand};
 use anyhow::{bail, Result};
 use clap::{ArgGroup, Parser, ValueHint};
 use pineappl::bin::BinRemapper;
+use std::ops::RangeInclusive;
 use std::path::PathBuf;
 
 /// Sums two or more bins of a grid together.
@@ -17,13 +18,22 @@ pub struct Opts {
     /// Sums all bins into a single bin.
     #[clap(long, group = "mode")]
     integrated: bool,
+    /// Merge specific bins together.
+    #[clap(
+        long,
+        group = "mode",
+        parse(try_from_str = helpers::try_parse_integer_range),
+        short,
+        use_value_delimiter = true,
+    )]
+    bins: Option<RangeInclusive<usize>>,
 }
 
 impl Subcommand for Opts {
     fn run(&self) -> Result<()> {
-        if self.integrated {
-            let mut grid = helpers::read_grid(&self.input)?;
+        let mut grid = helpers::read_grid(&self.input)?;
 
+        if self.integrated {
             if grid.merge_bins(0..grid.bin_info().bins()).is_err() {
                 bail!("TODO");
             }
@@ -41,11 +51,15 @@ impl Subcommand for Opts {
             key_values.remove("y_label");
             key_values.remove("y_label_tex");
             key_values.remove("y_unit");
-
-            helpers::write_grid(&self.output, &grid)
+        } else if let Some(range) = self.bins.as_ref() {
+            if grid.merge_bins(*range.start()..(range.end() + 1)).is_err() {
+                bail!("TODO");
+            }
         } else {
             unreachable!();
         }
+
+        helpers::write_grid(&self.output, &grid)
     }
 }
 
@@ -58,15 +72,27 @@ mod tests {
 Sums two or more bins of a grid together
 
 USAGE:
-    pineappl sum <--integrated> <INPUT> <OUTPUT>
+    pineappl sum <--integrated|--bins <BINS>> <INPUT> <OUTPUT>
 
 ARGS:
     <INPUT>     Path to the input grid
     <OUTPUT>    Path to the modified PineAPPL file
 
 OPTIONS:
-    -h, --help          Print help information
-        --integrated    Sums all bins into a single bin
+    -b, --bins <BINS>    Merge specific bins together
+    -h, --help           Print help information
+        --integrated     Sums all bins into a single bin
+";
+
+    const BINS_STR: &str = "bin   etal    disg/detal  scale uncertainty
+---+----+----+-----------+--------+--------
+  0    2 2.25 3.7527620e2   -3.77%    2.71%
+  1 2.25  2.5 3.4521553e2   -3.79%    2.80%
+  2  2.5 2.75 3.0001406e2   -3.78%    2.86%
+  3 2.75    3 2.4257663e2   -3.77%    2.92%
+  4    3 3.25 1.8093343e2   -3.74%    2.95%
+  5 3.25  3.5 1.2291115e2   -3.71%    2.98%
+  6  3.5  4.5 3.5811524e1   -3.59%    2.95%
 ";
 
     #[test]
@@ -77,6 +103,35 @@ OPTIONS:
             .assert()
             .success()
             .stdout(HELP_STR);
+    }
+
+    #[test]
+    fn bins() {
+        let output = NamedTempFile::new("bins.pineappl.lz4").unwrap();
+
+        Command::cargo_bin("pineappl")
+            .unwrap()
+            .args(&[
+                "sum",
+                "--bins=6-7",
+                "data/LHCB_WP_7TEV.pineappl.lz4",
+                output.path().to_str().unwrap(),
+            ])
+            .assert()
+            .success()
+            .stdout("");
+
+        Command::cargo_bin("pineappl")
+            .unwrap()
+            .args(&[
+                "--silence-lhapdf",
+                "convolute",
+                output.path().to_str().unwrap(),
+                "NNPDF31_nlo_as_0118_luxqed",
+            ])
+            .assert()
+            .success()
+            .stdout(BINS_STR);
     }
 
     #[test]